% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{change_to_name}
\alias{change_to_name}
\title{Change Column Name in Children to "name"}
\usage{
change_to_name(x, column = 1)
}
\arguments{
\item{x}{\code{data.frame} or \code{data.frame} derivative, such
as \code{tibble}}

\item{column}{column to convert}
}
\value{
\code{data.frame}
}
\description{
Change Column Name in Children to "name"
}
