% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{d3_dep_v3}
\alias{d3_dep_v3}
\title{'d3.js' Dependency for Version 3}
\usage{
d3_dep_v3()
}
\value{
\code{htmltools::htmlDependency}
}
\description{
'd3.js' Dependency for Version 3
}
\examples{
library(d3r)
library(htmltools)

attachDependencies(tagList(),d3_dep_v3())
}

