% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{d3_dep_v4}
\alias{d3_dep_v4}
\title{'d3.js' Dependency for Version 4}
\usage{
d3_dep_v4()
}
\value{
\code{htmltools::htmlDependency}
}
\description{
'd3.js' Dependency for Version 4
}
\examples{
library(d3r)
library(htmltools)

attachDependencies(tagList(),d3_dep_v4())
}

