% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{d3_dep_v6}
\alias{d3_dep_v6}
\title{'d3.js' Dependency for Version 6}
\usage{
d3_dep_v6(offline = TRUE)
}
\arguments{
\item{offline}{\code{logical} to specify whether to use
a local copy of d3.js (\code{TRUE}) or use cdn (\code{FALSE})}
}
\value{
\code{htmltools::htmlDependency}
}
\description{
'd3.js' Dependency for Version 6
}
\examples{
library(d3r)
library(htmltools)

tagList(d3_dep_v6())
}
\seealso{
\code{\link{d3_dep_v5}}, \code{\link{d3_dep_v4}}, \code{\link{d3_dep_v3}},
  and \code{\link{d3_dep_jetpack}}.

Other 'd3' dependency functions: 
\code{\link{d3_dep_jetpack}()},
\code{\link{d3_dep_v3}()},
\code{\link{d3_dep_v4}()},
\code{\link{d3_dep_v5}()}
}
\concept{'d3' dependency functions}
