% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{d3_nest}
\alias{d3_nest}
\title{Convert a \code{data.frame} to a 'd3.js' Hierarchy}
\usage{
d3_nest(data = NULL, value_cols = character(), root = "root")
}
\arguments{
\item{data}{\code{data.frame} or \code{data.frame} derivative, such
as \code{tibble}}

\item{value_cols}{\code{character} vector with the names of the
columns to use as data}

\item{root}{\code{character} name of the root level of the hierarchy}
}
\value{
nested \code{data.frame}
}
\description{
Convert a \code{data.frame} to a 'd3.js' Hierarchy
}
\examples{
# convert Titanic to a nested d3 hierarchy

# devtools::install_github("timelyportfolio/d3r")
library(d3r)
library(dplyr)

titanic_df <- data.frame(Titanic)
tit_tb <- titanic_df \%>\%
  select(Class,Age,Survived,Sex,Freq) \%>\%
  d3_nest(value_cols="Freq", root="titanic")

tit_tb

# see as json
jsonlite::toJSON(tit_tb, dataframe="row", auto_unbox=TRUE)
}

