% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{d3_table}
\alias{d3_table}
\title{Converts Table to 'd3' Nodes and Links}
\usage{
d3_table(tB = NULL, vars = NULL, agg = "Freq")
}
\arguments{
\item{tB}{\code{table} to convert}

\item{vars}{\code{character} vector of column names}

\item{agg}{\code{character} column name of aggregated value}
}
\value{
\code{list} of two \code{data.frames} - one for nodes and
          one for links of the network.  This structure is helpful
          when working with \code{networkD3} and \code{visNetwork}.
}
\description{
Converts Table to 'd3' Nodes and Links
}
\examples{
library(d3r)
d3_table(Titanic, c("Class","Sex"))

}
