% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8.R
\name{d3_v8}
\alias{d3_v8}
\title{Create V8 Context with D3}
\usage{
d3_v8(...)
}
\arguments{
\item{...}{arguments passed to \code{v8()}}
}
\value{
\code{v8} context with d3.js sourced
}
\description{
Create V8 Context with D3
}

