% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{myD3Visualizer}
\alias{myD3Visualizer}
\title{Create a d3 Visualizer Widget}
\usage{
myD3Visualizer(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A numeric vector for the circle radii.}

\item{width}{The width of the widget (optional).}

\item{height}{The height of the widget (optional).}

\item{elementId}{An element id for the widget's container (optional).}
}
\value{
An htmlwidget object.
}
\description{
This widget uses d3.js to render a series of circles where each circle's radius is given
by the numeric data provided.
}
