% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rstudio.R
\name{prepare_rstudio}
\alias{prepare_rstudio}
\title{Load a Matrix}
\usage{
prepare_rstudio(html)
}
\arguments{
\item{html}{HTML string}
}
\value{
HTML displayed in RStudio Viewer
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
