% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{promote_na}
\alias{promote_na}
\title{Apply `promote_na` to All Rows}
\usage{
promote_na(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
Apply `promote_na` to All Rows
}

