% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{promote_na_one}
\alias{promote_na_one}
\title{Promote NA to Top Level}
\usage{
promote_na_one(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
Promote NA to Top Level
}
