% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_html.R
\name{serve_html}
\alias{serve_html}
\title{Serve HTML}
\usage{
serve_html(html, dir)
}
\arguments{
\item{html}{HTML string}

\item{dir}{Directory for HTML file}
}
\value{
File \code{index.html} in \code{dir} directory
}
\description{
This function saves D3.js output in a folder
}
