% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_rstudio.R
\name{serve_rstudio}
\alias{serve_rstudio}
\title{Serve D3.js in RStudio}
\usage{
serve_rstudio(html)
}
\arguments{
\item{html}{HTML string}
}
\value{
HTML displayed in RStudio Viewer
}
\description{
This function serves D3.js output in RStudio Viewer
}
