% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_shiny.R
\name{serve_shiny}
\alias{serve_shiny}
\title{Serve D3.js in Shiny}
\usage{
serve_shiny(html, width, height)
}
\arguments{
\item{html}{HTML string}

\item{width}{iframe width}

\item{height}{iframe height}
}
\value{
Shiny HTML tag iframe
}
\description{
This function serves D3.js output in Shiny
}
