% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{d3r}
\alias{d3r}
\title{d3r: R6 Class for D3.js}
\description{
This class serves D3.js output in R ecosystem.
}
\examples{
# Usage Example:
chart <- d3r$new(
  script = 'console.log(`Hello, ${data[0]}`);',
  title = "Check the Browser Console (F12 → Console Tab)",
  mode = NULL,
  data = jsonlite::toJSON(Sys.info()[["user"]])
  )
chart$make()
file.remove("index.html")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{script}}{JavaScript string}

\item{\code{style}}{(optional) CSS string}

\item{\code{html}}{(optional) HTML template string}

\item{\code{data}}{(optional) JSON string. This adds line \verb{const data = <your_data>;} to the JavaScript.}

\item{\code{title}}{(optional) Title}

\item{\code{mode}}{Available modes: NULL, Jupyter, RStudio, Quarto, Shiny, Selenium}

\item{\code{dir}}{(optional) Directory to store HTML/SVG file}

\item{\code{width}}{(optional) iframe width for Shiny and Quarto iframe}

\item{\code{height}}{(optional) iframe height for Shiny and Quarto iframe}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-d3r-new}{\code{d3r$new()}}
\item \href{#method-d3r-generate_html}{\code{d3r$generate_html()}}
\item \href{#method-d3r-make}{\code{d3r$make()}}
\item \href{#method-d3r-clone}{\code{d3r$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d3r-new"></a>}}
\if{latex}{\out{\hypertarget{method-d3r-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class with default values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d3r$new(
  script,
  style = NULL,
  html = NULL,
  data = NULL,
  title = NULL,
  mode = NULL,
  dir = NULL,
  width = 600,
  height = 400
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{script}}{JavaScript string}

\item{\code{style}}{(optional) CSS string}

\item{\code{html}}{(optional) HTML template string}

\item{\code{data}}{(optional) JSON string. This adds line \verb{const data = <your_data>;} to the JavaScript.}

\item{\code{title}}{(optional) Title}

\item{\code{mode}}{Available modes: NULL, Jupyter, RStudio, Quarto, Shiny, Selenium}

\item{\code{dir}}{(optional) Directory to store HTML/SVG file}

\item{\code{width}}{(optional) iframe width for Shiny and Quarto iframe}

\item{\code{height}}{(optional) iframe height for Shiny and Quarto iframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d3r-generate_html"></a>}}
\if{latex}{\out{\hypertarget{method-d3r-generate_html}{}}}
\subsection{Method \code{generate_html()}}{
generate_html the class with default values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d3r$generate_html()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d3r-make"></a>}}
\if{latex}{\out{\hypertarget{method-d3r-make}{}}}
\subsection{Method \code{make()}}{
make the class with default values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d3r$make()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-d3r-clone"></a>}}
\if{latex}{\out{\hypertarget{method-d3r-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{d3r$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
