% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_quarto.R
\name{serve_quarto}
\alias{serve_quarto}
\title{Serve D3.js in Quarto}
\usage{
serve_quarto(html, dir, width, height)
}
\arguments{
\item{html}{HTML string}

\item{dir}{Directory for HTML file}

\item{width}{iframe width}

\item{height}{iframe height}
}
\value{
Iframe embedded in Quarto or R Markdown document
}
\description{
This function serves D3.js output in Quarto
}
