% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_selenium.R
\name{serve_selenium}
\alias{serve_selenium}
\title{Serve D3.js in Selenium}
\usage{
serve_selenium(html)
}
\arguments{
\item{html}{HTML string}
}
\value{
path of the screenshot PNG file
}
\description{
This function serves D3.js output in Selenium and takes screenshot
}
